refactor(cli): make the CLI runnable without re-importing its entrypoint
CI failed intermittently with "Test timed out in 5000ms" across
tests/unit/cli. Not a hang and not a race: bootstrap.ts invoked
runBootstrap() at module load, so any test wanting a fresh run had to
re-import it through vi.resetModules(). A registry reset does not clear
vitest's transform cache, so whichever case imported first absorbed a
cold transform of the whole command graph inside its own 5s budget —
measured 1.4-2.4s idle and up to 4.2s under saturating CPU load, i.e.
85% of the budget spent before the suite's own workers added contention.
That is why expansion-help.test.ts was the 8th slowest file in the suite
while holding only two tests.

The logic moves to run.ts as an exported runCli(); bootstrap.ts stays the
esbuild entrypoint and does nothing but invoke it. Keeping the invocation
out of the logic module is the whole point, so both files say so. The two
tests that only needed freshness now use hoisted vi.mock plus a static
import and call runCli() per case: 1775ms and 2405ms became 2ms and 6ms,
and 7ms and 27ms under saturating load.

Three more files hit the same wall for a different reason — they swap
module doubles per case, so their registry resets are load-bearing and
the re-execution cost is inherent. They warm the transform cache once in
beforeAll instead, which halves their worst case from ~1.4s to ~0.75s.

That residual is real, and more files carry the same latent pattern (38
use resetModules), so the suite now states a deliberate testTimeout
rather than inheriting vitest's 5s by accident. 15s is honest for a suite
whose own config documents it as I/O-bound with workers oversubscribed,
and still fails a genuinely hung test promptly.

Verified: 4314/4314 across six consecutive runs, all five affected files
green under 32-way CPU saturation, and the built bundle's entry exercised
end-to-end via the store-reset build contract.
