+++ tests/unit/cli/coral-store-read-parity.test.ts
@@ -4,7 +4,7 @@ import { join } from 'node:path';
 import { tmpdir } from 'node:os';
 
 import type { Command } from 'commander';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 import type * as MainMod from '#src/cli/program.js';
 
 import { pluginRootNamespace } from '#src/infra/plugin-identity.js';
@@ -134,6 +134,13 @@ async function runCli(program: Command, args: string[]): Promise<{ stdout: strin
 }
 
 describe('cli coral-store read parity', () => {
+  // `loadMainModule()` resets the module registry per case because these cases swap module doubles, but a
+  // registry reset does not clear vitest's transform cache. Warm it once here so the first case does not
+  // absorb ~1.4s of cold command-graph transform inside its 5s budget, which flaked under CI contention.
+  beforeAll(async () => {
+    await import('#src/cli/program.js');
+  });
+
   let tempHome: string;
   let projectRoot: string;
   let originalHome: string | undefined;
