+++ tests/unit/cli/expansion-help.test.ts
@@ -1,27 +1,26 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
+import { runCli } from '#src/cli/run.js';
 import { installErrorSchema } from '#src/expansion/rpc-contract.js';
 
 function toText(chunk: string | Uint8Array): string {
   return typeof chunk === 'string' ? chunk : Buffer.from(chunk).toString('utf8');
 }
 
-// Each case re-imports `bootstrap.ts` after `vi.resetModules()` (it runs on load) and deliberately uses the
-// real `buildProgram()`, so the first case absorbs a cold transform+resolve of the whole CLI command graph
-// (~1.2s idle). `vitest/default.ts` oversubscribes workers on CI (4 workers on a 2-core runner) by design, so
-// that inflates under contention past the 5s default. Budget for the import; a genuine hang still trips this.
-describe('expansion bootstrap output', { timeout: 30_000 }, () => {
+// Static import of `runCli`, invoked per case. Re-importing a self-executing entry through
+// `vi.resetModules()` used to charge this file's cold transform of the real command graph to whichever case
+// ran first, which flaked against the 5s default; collection absorbs it now.
+describe('expansion bootstrap output', () => {
   const originalArgv = [...process.argv];
   let stdout = '';
   let stderr = '';
 
-  async function runBootstrap(argv: string[]): Promise<void> {
+  async function run(argv: string[]): Promise<void> {
     stdout = '';
     stderr = '';
     process.argv = ['node', 'coral-cli', ...argv];
     process.exitCode = undefined;
 
-    vi.resetModules();
     vi.spyOn(process, 'exit').mockImplementation(((_code?: number) => undefined) as typeof process.exit);
     vi.spyOn(process.stdout, 'write').mockImplementation(((chunk: string | Uint8Array) => {
       stdout += toText(chunk);
@@ -32,19 +31,17 @@ describe('expansion bootstrap output', { timeout: 30_000 }, () => {
       return true;
     }) as typeof process.stderr.write);
 
-    const { bootstrapCompletion } = await import('#src/cli/bootstrap.js');
-    await bootstrapCompletion;
+    await runCli();
   }
 
   afterEach(() => {
     process.argv = [...originalArgv];
     process.exitCode = undefined;
     vi.restoreAllMocks();
-    vi.resetModules();
   });
 
   it('keeps expansion --help visible on stdout with exit 0', async () => {
-    await runBootstrap(['expansion', '--help']);
+    await run(['expansion', '--help']);
 
     expect(stderr).toBe('');
     expect(stdout.trim().length).toBeGreaterThan(0);
@@ -55,7 +52,7 @@ describe('expansion bootstrap output', { timeout: 30_000 }, () => {
   });
 
   it('normalizes expansion equip missing-argument failures to one stdout InstallError JSON line', async () => {
-    await runBootstrap(['expansion', 'equip']);
+    await run(['expansion', 'equip']);
 
     expect(stderr).toBe('');
     expect(stdout.endsWith('\n')).toBe(true);
