+++ tests/unit/cli/main-routing.test.ts
@@ -3,7 +3,7 @@ import { mkdtempSync, rmSync, writeFileSync, unlinkSync } from 'node:fs';
 import { join } from 'node:path';
 import { tmpdir } from 'node:os';
 import type { Command } from 'commander';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 import type * as CommandClientMod from '#src/cli/dispatch.js';
 import type * as CommandOutputMod from '#src/cli/emit.js';
 import type * as ErrorsMod from '#src/cli/errors.js';
@@ -338,6 +338,13 @@ function createCauseRenderFixture(): { home: string; cleanup(): void } {
 }
 
 describe('cli main routing', () => {
+  // `loadMainModule()` resets the module registry per case because these cases swap module doubles, but a
+  // registry reset does not clear vitest's transform cache. Warm it once here so the first case does not
+  // absorb ~1.4s of cold command-graph transform inside its 5s budget, which flaked under CI contention.
+  beforeAll(async () => {
+    await import('#src/cli/program.js');
+  });
+
   let stdout = '';
   let stderr = '';
   let originalArgv: string[];
